requirements.txt
mecab-python3
mypy
pylint
google-cloud-texttospeech